fix(savePos): Fixing weird savePos bug
During a live session yeserday I noticed an issue where a user had used
`class` as a param name. Since it's a keyword, prism turned that into a
nested span.

This is a pretty rare event, and so it caused savePos() to calculate the
caretPos inaccurately which meant the delete/insert operation was
happening in the wrong place.

This commit fixes that.